backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.8
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
